# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Employee API
  description: REST API for managing employees, projects, and tasks. The API uses
    HTTPS and JSON responses with consistent error handling. Task creation may trigger
    asynchronous notification processing through a background worker.
  version: 1.1.0
servers:
- url: https://localhost:8443
  description: Local HTTPS server
...
paths:
...
  /api/v1/employees:
    get:
      operationId: listEmployees
      tags:
      - Employees
      summary: Get employees
      description: Get a paginated list of employees with optional filtering and sorting.
        Use '-field' for descending order.
      parameters:
      - name: page
        in: query
        required: false
        description: Page number.
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: page_size
        in: query
        required: false
        description: Number of employees per page.
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
      - name: name
        in: query
        required: false
        description: Filter employees by name.
        schema:
          type: string
      - name: email
        in: query
        required: false
        description: Filter employees by email.
        schema:
          type: string
      - name: sort
        in: query
        required: false
        description: Sort field. Prefix with '-' for descending order.
        schema:
          type: string
          enum:
          - id
          - -id
          - name
          - -name
          - email
          - -email
          default: id
          example: -name
      responses:
        '200':
          description: Employees retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createEmployee
      tags:
      - Employees
      summary: Create employee
      description: Create a new employee.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          description: Invalid employee data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/employees/{id}:
    get:
      operationId: getEmployee
      tags:
      - Employees
      summary: Get employee
      description: Get a single employee by ID.
      parameters:
      - name: id
        in: path
        required: true
        description: Employee ID.
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Employee retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          description: Invalid employee ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateEmployee
      tags:
      - Employees
      summary: Update employee
      description: Update an existing employee.
      parameters:
      - name: id
        in: path
        required: true
        description: Employee ID.
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Employee updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          description: Invalid employee data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteEmployee
      tags:
      - Employees
      summary: Delete employee
      description: Delete an employee by ID.
      parameters:
      - name: id
        in: path
        required: true
        description: Employee ID.
        schema:
          type: integer
          minimum: 1
      responses:
        '204':
          description: Employee deleted successfully.
        '400':
          description: Invalid employee ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects:
    get:
      operationId: listProjects
      tags:
      - Projects
      summary: Get all projects
      description: Get a list of all projects.
      responses:
        '200':
          description: Projects retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createProject
      tags:
      - Projects
      summary: Create project
      description: Create a new project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid project data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{id}:
    get:
      operationId: getProject
      tags:
      - Projects
      summary: Get project
      description: Get a single project by ID.
      parameters:
      - name: id
        in: path
        required: true
        description: Project ID.
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Project retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid project ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateProject
      tags:
      - Projects
      summary: Update project
      description: Update an existing project.
      parameters:
      - name: id
        in: path
        required: true
        description: Project ID.
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid project data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteProject
      tags:
      - Projects
      summary: Delete project
      description: Delete a project by ID.
      parameters:
      - name: id
        in: path
        required: true
        description: Project ID.
        schema:
          type: integer
          minimum: 1
      responses:
        '204':
          description: Project deleted successfully.
        '400':
          description: Invalid project ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks:
    get:
      operationId: listTasks
      tags:
      - Tasks
      summary: Get all tasks
      description: Get a list of all tasks.
      responses:
        '200':
          description: Tasks retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createTask
      tags:
      - Tasks
      summary: Create task
      description: Create a new task. After successful creation, a task.created event
        may be published for asynchronous notification processing by a background
        worker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid task data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks/{id}:
    get:
      operationId: getTask
      tags:
      - Tasks
      summary: Get task
      description: Get a single task by ID.
      parameters:
      - name: id
        in: path
        required: true
        description: Task ID.
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Task retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid task ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateTask
      tags:
      - Tasks
      summary: Update task
      description: Update an existing task.
      parameters:
      - name: id
        in: path
        required: true
        description: Task ID.
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid task data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteTask
      tags:
      - Tasks
      summary: Delete task
      description: Delete a task by ID.
      parameters:
      - name: id
        in: path
        required: true
        description: Task ID.
        schema:
          type: integer
          minimum: 1
      responses:
        '204':
          description: Task deleted successfully.
        '400':
          description: Invalid task ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Not acceptable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateEmployeeRequest:
      type: object
      required:
      - name
      - email
      properties:
        name:
          type: string
          example: Ali
        email:
          type: string
          format: email
          example: ali@example.com
        department:
          type: string
          example: IT
    UpdateEmployeeRequest:
      type: object
      required:
      - name
      - email
      properties:
        name:
          type: string
          example: Ali Ahmed
        email:
          type: string
          format: email
          example: ali.ahmed@example.com
        department:
          type: string
          example: HR
    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Ali
        email:
          type: string
          format: email
          example: ali@example.com
        department:
          type: string
          example: IT
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          example: 10
        total:
          type: integer
          minimum: 0
          example: 3
        total_pages:
          type: integer
          minimum: 0
          example: 1
    EmployeeListResponse:
      type: object
      required:
      - data
      - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateProjectRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          example: Project Alpha
        description:
          type: string
          example: This is the first project
    UpdateProjectRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          example: Project Alpha Updated
        description:
          type: string
          example: Updated description
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Project Alpha
        description:
          type: string
          example: This is the first project
    CreateTaskRequest:
      type: object
      required:
      - project_id
      - title
      properties:
        project_id:
          type: integer
          minimum: 1
          example: 1
        title:
          type: string
          example: Design database schema
        status:
          type: string
          enum:
          - pending
          - in_progress
          - completed
          example: pending
          default: pending
    UpdateTaskRequest:
      type: object
      required:
      - project_id
      - title
      properties:
        project_id:
          type: integer
          minimum: 1
          example: 1
        title:
          type: string
          example: Design database schema - Updated
        status:
          type: string
          enum:
          - pending
          - in_progress
          - completed
          example: in_progress
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        project_id:
          type: integer
          minimum: 1
          example: 1
        title:
          type: string
          example: Design database schema
        status:
          type: string
          enum:
          - pending
          - in_progress
          - completed
          example: pending
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: EMPLOYEE_NOT_FOUND
            message:
              type: string
              example: employee not found
...